include tilt in temp unit config
